Add edge probability step and add to loss
Realized that the simplest method here is to only use the valid
molecular formula which is provided as a label to the Data() object.
We ignore the BOA prediction since it will have potentially missing and/or
extra atoms, and thus the loss calculation is difficult.

One downside of this approach is that we have to carry around an adjacency matrix
which is quite large since by definition it has to be fully connected.

Note that I have to double check that the output edge predictions are getting closer
to the expected result (the DEBUG flag that I added)