Add some debug functions but it seems that I might be doing something wrong
After initialization, we get that the edge most likely is a triple bond

Then very quickly the model learns to set the bond to 0 (no connection) and somehow
that turns down the loss. I think though that this is not real gains because
we're not focusing on the connections that are there, just the overall adjacency matrix

So I think the issue (apart from maybe an implementation issue/bug) is that we are still giving
a loss decrease when the model correctly predicts 'None' for all of the bonds, but the truth is that
the zero bond label is much less important than the labels where atoms ARE connected.

Before exploring this, I need to make sure I am setting everything up correctly (although
as a first pass it looks good).