Add graph_id to track Data object, and add positional encoding
By attaching a "graph_id" to each Data object, we can now map the object to the SMILES string which helps with debugging

Added positional encoding as per the paper. At the beginning of training, it seems that all the edge probability matrix can predict is that all the atoms are not connected to each other. However, as  I trained for maybe 50 epochs, I finally started getting bond predictions that were non-zero.

I think the positional encoding really helped because the bond edges all start off with the same feature vector. If all the node features are the same as well, then the model will not really learn to differentiate edges or nodes, thus since most of the nodes should be zero, the model will have difficulty guessing non-zero bonds for most of the atoms (how would it be able to tell?)

I explored this very briefly by looking at a couple dimensions of the 'e' feature vector after going through the GCN. It seems that at the beginning these embeddings are way too close together, and the node features do not carry enough weight.