Attempt VAE implementation
PyTorch implementation of a VAE, where VAE learns a parameterization
of the molecular latent vector z as the sum of a mean vector and a std
vector which is multiplied by a sampled normal gaussian.

The KL loss steadily decreases, and the shapes match up, so I think that the
implementation is correct. To verify though, we should set the embedding dimension to 2 and then plot the
encoded z of each training sample. Visually they should cluster to form approximately a gaussian.

Alternatively, we can just directly try sampling from a gaussian and decoding it to
see what outputs we obtain.