Add jupyter notebook
The jupyter notebook integrates pretty much all the code into one location. This jupyter notebook content will be converted into an interactive streamlit app.

The reason I'm using streamlit is to have an interactive application... it will be hosted on Hugging Face Spaces and be embedded as an iframe on my personal website.